backend/requirements.txt
# Django y DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Autenticación JWT
djangorestframework-simplejwt==5.3.0

# Base de datos
psycopg2-binary==2.9.7
dj-database-url==3.0.1
django-redis==5.4.0

# Cache y sesiones
redis==5.0.1

# Validación y serialización
drf-spectacular==0.26.5
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Linting y formateo
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Seguridad
django-ratelimit==4.1.0
django-csp==3.7

# Utilidades
python-decouple==3.8
Pillow==10.1.0
celery==5.3.4

# Documentación API
drf-yasg==1.21.7

# Variables de entorno
python-dotenv==1.0.0

# Para desarrollo
django-debug-toolbar==4.2.0 